+++ NEWS.md
@@ -10,6 +10,7 @@
 - S3 method: crystal.pdb;
 - Function lattice.range (as alias for range.lattice.pdb);
 - Function mark.pdb: but still hidden;
+- Function rm.h: remove hydrogen (for NMR structures) - still hidden;
 
 ## Enhancement:
 - [fix] Model extraction;